WIP: keymgmt: move key management to separate module
The key management is moved to a separate module and its
implementation has been adapted to the key objects. The key management
is still not fully functional, but loading a EVP_PKEY by URI is
working now.

The move causes some problems in the provider module, which are
commented out for the moment. The changes in the provider module will
be reworked and reverted respectively.

Fix compile warning in provider module.

Signed-off-by: Holger Dengler <dengler@linux.ibm.com>