WIP: pkcs11: add helper for deep operations on attributes
CK_ATTRIBUTEs hold pointers to memory on the heap, which is controlled
by the PKCS#11 module. If an CK_ATTRIBUTE needs to be duplicated, a
memdup() is not sufficient, the content of pValue must also copied. To
free an CK_ATTRIBUTE, its pValues must be freed.  Add helper functions
for deep memdup and free.

Signed-off-by: Holger Dengler <dengler@linux.ibm.com>