signature: rework parameter handling
The new parameter handling reduces the complexity of the PKCS#11
mechanism handling. The operation context of the forward provider is
used to manage all parameters. The PKCS#11 mechanism is setup right
before the PKCS#11 operations. Also complete support for padding
modes.

Ensure the PKCS#11 session and object handles before processing the
PKCS#11 operations.

Suggested-by: Ingo Franzki <ifranzki@linux.ibm.com>
Signed-off-by: Holger Dengler <dengler@linux.ibm.com>