pkcs11: Allow CKR_CRYPTOKI_ALREADY_INITIALIZED with C_Initialize
The provider initializes the PKCS#11 library using C_Initialize during
provider initialization. If this fails, then the provider fails to
initialize and is not available.

There are situations where applications first initialize the PKCS#11
library themselves, and then load the provider, e.g. due to explicit or
implicit OpenSSL initialization which loads the provider. In such cases
the provider's C_Initialize will fail with CKR_CRYPTOKI_ALREADY_INITIALIZED.

Ignore this error situation, because the PKCS#11 library is already
initialized, and thus the provider can simply continue and call further
PKCS#11 services. However, in such case, the provider should also not
call C_Finalize during its teardown, but leave it to the application to
finalize the PKCS#11 library.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>