keymgmt: Fail to export a private key when it is a PKCS#11 key.
As per https://www.openssl.org/docs/man3.0/man7/crypto.html section
"Implicit fetching" under "ALGORITHM FETCHING", OpenSSL may try to export
the key if an operation is to be performed that is implemented by another
provider. However, if the export fails, then it tries to fetch the operation
from the same provider as the key.

We can not export the private key parts of a PKCS#11 based key, so fail
in case the selection includes the private key. OpenSSL will then fetch
the operation (i.e. signature or asym-cipher) from our provider and perform
the operation that way.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>