store: move object retrieval to store_load()
Move PKCS#11 object retrieval from store_open() to store_load(). This
is a preparation for using the passphrase call-back in store_load() to
acquire the token PIN from the caller.

As a workaround, store_eof() call will return false until the number
of available keys is retrieved at the first store_load() call.

Signed-off-by: Holger Dengler <dengler@linux.ibm.com>