store: implement store_open_ex() interface
The current store_open() interface does not get a pointer to the
passphrase call-back, which makes it impossible to get the number of
available keys during the store_open call, if no pin is specified in
the URI. As a workaround, the number of available keys is retrieved at
the first store_load() call.

OpenSSL 3.2 introduced the new store_open_ex() interface with
additional arguments, e.g. a pointer to the passphrase
call-back. Implement this new interface if it is available in the
OpenSSL library.

Fixes: #5
Signed-off-by: Holger Dengler <dengler@linux.ibm.com>